:bug: Solving circular dependecies and string resolution in SQLAlchemy mapper
By the module-functional structure of the project a problem showed up in
the models, in the database context the entities have relationships and
Python had some troubles with the depedencies resulting in circular ones
and also SQLAlchemy had problems with the string resolutions of Model
names in the mapper. This commits fixed those issues.