:passport_control: Update login and signup logic
The application provides a feature to connect with google account and
for system designs reasons it would be complicated to have separate
system for authentication (traditional way) and at the same time use
google flow, credentials, etc. and also in the end for the end-user it
would be much easier just to log in the system with google account as to
create a separate account and then also need to connect with google. It
also isn't good from the programmers perspective in maintaining the
code.