:sparkles: Add workflow router (wrapping the workflow features with FastAPI)
This commit adds the workflow router with the get workflows, toggle
on/off workflow, and update parameter workflow features that operate in
the database layer through Workflow Service and also in the
orchestration layer through Deployment Service.

And because of sequential dependencies between the workflow and
deployment services the session changes are commited in the router.