:bug: Fixed possible bug
The connected_account from the AccountService.get_account() method is a
SQLAlchemy model so if we try to access the model outside the with
statement where the database session is closed it can cause errors. So
to prevent that we indent the line into the 'with' statment to save the
history_id.